+++ http_client.go
@@ -9,35 +9,35 @@ import (
 	"time"
 )
 
-type API struct {
+type HTTP struct {
 	BaseUrl string
 	client  *http.Client
 }
 
-func NewAPI(baseUrl string) *API {
-	return &API{
+func NewHTTP(baseUrl string) *HTTP {
+	return &HTTP{
 		BaseUrl: baseUrl,
 		client:  http.DefaultClient,
 	}
 }
 
-func (a *API) Get(path string) (int, []byte) {
+func (a *HTTP) Get(path string) (int, []byte) {
 	return a.RetryRequest("GET", path, []byte{})
 }
 
-func (a *API) Post(path string, data []byte) (int, []byte) {
+func (a *HTTP) Post(path string, data []byte) (int, []byte) {
 	return a.RetryRequest("POST", path, data)
 }
 
-func (a *API) Put(path string, data []byte) (int, []byte) {
+func (a *HTTP) Put(path string, data []byte) (int, []byte) {
 	return a.RetryRequest("PUT", path, data)
 }
 
-func (a *API) Delete(path string) (int, []byte) {
+func (a *HTTP) Delete(path string) (int, []byte) {
 	return a.RetryRequest("DELETE", path, []byte{})
 }
 
-func (a *API) RetryRequest(method string, path string, data []byte) (int, []byte) {
+func (a *HTTP) RetryRequest(method string, path string, data []byte) (int, []byte) {
 	for x := 0; x < 5; x++ {
 		i, b, err := a.Request(method, path, data)
 		if err == nil {
@@ -50,7 +50,7 @@ func (a *API) RetryRequest(method string, path string, data []byte) (int, []byte
 	return 0, nil
 }
 
-func (a *API) Request(method string, path string, data []byte) (int, []byte, error) {
+func (a *HTTP) Request(method string, path string, data []byte) (int, []byte, error) {
 	url := fmt.Sprintf("%s%s", a.BaseUrl, path)
 	req, err := http.NewRequest(method, url, bytes.NewBuffer(data))
 	if err != nil {
